+++ tests/utils/test_fp4_quantize.py
@@ -15,7 +15,19 @@ from flashinfer import (
     silu_and_mul_scaled_nvfp4_experts_quantize,
     silu_and_mul,
 )
-from flashinfer.utils import is_sm100a_supported
+
+
+def is_fp4_supported(device: torch.device) -> bool:
+    """Check if FP4 quantization is supported (SM100+ with CUDA >= 12.8)."""
+    major, _ = torch.cuda.get_device_capability(device)
+    cuda_version = torch.version.cuda
+    if cuda_version is None:
+        return False
+    # Check CUDA >= 12.8
+    parts = cuda_version.split(".")
+    cuda_major, cuda_minor = int(parts[0]), int(parts[1]) if len(parts) > 1 else 0
+    return major >= 10 and (cuda_major > 12 or (cuda_major == 12 and cuda_minor >= 8))
+
 
 DTYPES = [torch.float16, torch.bfloat16]
 # The batch dimension doesn't need to be multiple of 128
@@ -112,8 +124,8 @@ def test_fp4_quantization(
     sf_use_ue8m0: bool,
     is_swizzled: bool,
 ) -> None:
-    if not is_sm100a_supported(torch.device(device)):
-        pytest.skip("Nvfp4 Requires compute capability >= 10 and CUDA >= 12.8")
+    if not is_fp4_supported(torch.device(device)):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
     m, n = shape
@@ -155,8 +167,8 @@ def test_scale_swizzling(
     seed: int,
     device: str,
 ) -> None:
-    if not is_sm100a_supported(torch.device("cuda")):
-        pytest.skip("Nvfp4 Requires compute capability >= 10 and CUDA >= 12.8")
+    if not is_fp4_supported(torch.device("cuda")):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
     m, n = shape
@@ -191,8 +203,8 @@ def test_block_scale_interleave(
     device: str,
 ) -> None:
     """Test the block_scale_interleave function directly."""
-    if not is_sm100a_supported(torch.device("cuda")):
-        pytest.skip("Nvfp4 Requires compute capability >= 10 and CUDA >= 12.8")
+    if not is_fp4_supported(torch.device("cuda")):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
 
@@ -240,8 +252,8 @@ def test_e2m1_dequantization(
     sf_use_ue8m0: bool,
 ) -> None:
     """Test roundtrip: fp4_quantize -> e2m1_and_ufp8sf_scale_to_float."""
-    if not is_sm100a_supported(torch.device("cuda")):
-        pytest.skip("Nvfp4 Requires compute capability >= 10 and CUDA >= 12.8")
+    if not is_fp4_supported(torch.device("cuda")):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
 
@@ -333,8 +345,8 @@ def test_mxfp4_quantize_roundtrip(
     device: str,
 ) -> None:
     """Test MXFP4 quantization roundtrip for both backends."""
-    if not is_sm100a_supported(torch.device(device)):
-        pytest.skip("MXFP4 Requires compute capability >= 10 and CUDA >= 12.8")
+    if not is_fp4_supported(torch.device(device)):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     if backend == "cute-dsl" and not _is_cute_dsl_available():
         pytest.skip("CuTe-DSL not available")
 
@@ -381,8 +393,8 @@ def test_mxfp4_quantize_backend_parity(
     device: str,
 ) -> None:
     """Test that CUDA and CuTe-DSL backends produce matching results."""
-    if not is_sm100a_supported(torch.device(device)):
-        pytest.skip("MXFP4 Requires compute capability >= 10 and CUDA >= 12.8")
+    if not is_fp4_supported(torch.device(device)):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     if not _is_cute_dsl_available():
         pytest.skip("CuTe-DSL not available")
 
@@ -463,8 +475,8 @@ def test_nvfp4_batched_quantize(
     device: str,
 ) -> None:
     """Test nvfp4_batched_quantize function."""
-    if not is_sm100a_supported(torch.device(device)):
-        pytest.skip("Nvfp4 Requires compute capability of 10 or above")
+    if not is_fp4_supported(torch.device(device)):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
 
@@ -506,8 +518,8 @@ def test_scaled_fp4_grouped_quantize(
     device: str,
 ) -> None:
     """Test scaled_fp4_grouped_quantize function."""
-    if not is_sm100a_supported(torch.device(device)):
-        pytest.skip("Nvfp4 Requires compute capability of 10 or above")
+    if not is_fp4_supported(torch.device(device)):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
 
@@ -553,8 +565,8 @@ def test_silu_and_mul_scaled_nvfp4_experts_quantize(
     device: str,
 ) -> None:
     """Test silu_and_mul_nvfp4_batched_quantize function."""
-    if not is_sm100a_supported(torch.device(device)):
-        pytest.skip("Nvfp4 Requires compute capability of 10 or above")
+    if not is_fp4_supported(torch.device(device)):
+        pytest.skip("FP4 requires SM100+ and CUDA >= 12.8")
     torch.set_default_device(device)
     torch.manual_seed(seed)
